+++ src/rules/lt.js
@@ -1,0 +1,19 @@
+const empty = require('../lib/empty');
+const { pathIndex } = require('../lib/ObjectIndex');
+const numeric = require('./numeric');
+
+module.exports = async function lt(field, value, anotherField) {
+
+
+    const anotherFieldValue = pathIndex(this.inputs, anotherField);
+
+    if (!(await numeric(field, anotherFieldValue))) {
+        return false;
+    }
+
+    if (Number(value) < Number(anotherFieldValue)) {
+        return true;
+    }
+
+    return false;
+}