+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- nothing
+- gt: greater then another field rle
+- gte: greater then or equals another field rle
+- lt: less then another field rle
+- lte: less then or equals another field rle
 
 ### Changed
 
@@ -22,6 +25,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - nothing
+
+### Fixed
+
+- nothing
+
+### Security
+
+- nothing
+
+## [3.3.0]
 
 ### Fixed